+++ src/etp/ServerInitializationParameters.cpp
@@ -35,32 +35,32 @@ std::vector<Energistics::Etp::v12::Datatypes::SupportedProtocol> ServerInitializ
 	protocolVersion.revision = 0;
 
 	Energistics::Etp::v12::Datatypes::SupportedProtocol protocol;
-	protocol.protocol = Energistics::Etp::v12::Datatypes::Protocol::Core;
+	protocol.protocol = static_cast<int32_t>(Energistics::Etp::v12::Datatypes::Protocol::Core);
 	protocol.protocolVersion = protocolVersion;
 	protocol.role = "server";
 	result.push_back(protocol);
 
-	protocol.protocol = Energistics::Etp::v12::Datatypes::Protocol::Discovery;
+	protocol.protocol = static_cast<int32_t>(Energistics::Etp::v12::Datatypes::Protocol::Discovery);
 	protocol.role = "store";
 	result.push_back(protocol);
 
-	protocol.protocol = Energistics::Etp::v12::Datatypes::Protocol::Store;
+	protocol.protocol = static_cast<int32_t>(Energistics::Etp::v12::Datatypes::Protocol::Store);
 	result.push_back(protocol);
 
-	protocol.protocol = Energistics::Etp::v12::Datatypes::Protocol::StoreNotification;
+	protocol.protocol = static_cast<int32_t>(Energistics::Etp::v12::Datatypes::Protocol::StoreNotification);
 	Energistics::Etp::v12::Datatypes::DataValue value;
 	value.item.set_long((std::numeric_limits<long>::max)());
 	protocol.protocolCapabilities["MaxDataArraySize"] = value;
 	result.push_back(protocol);
 	protocol.protocolCapabilities.clear();
 
-	protocol.protocol = Energistics::Etp::v12::Datatypes::Protocol::DataArray;
+	protocol.protocol = static_cast<int32_t>(Energistics::Etp::v12::Datatypes::Protocol::DataArray);
 	result.push_back(protocol);
 
-	protocol.protocol = Energistics::Etp::v12::Datatypes::Protocol::Transaction;
+	protocol.protocol = static_cast<int32_t>(Energistics::Etp::v12::Datatypes::Protocol::Transaction);
 	result.push_back(protocol);
 
-	protocol.protocol = Energistics::Etp::v12::Datatypes::Protocol::Dataspace;
+	protocol.protocol = static_cast<int32_t>(Energistics::Etp::v12::Datatypes::Protocol::Dataspace);
 	result.push_back(protocol);
 
 	return result;
